Refactor dataset code and fix bugs.
