+++ support_scripts/utils/__init__.py
@@ -1,8 +1,16 @@
 from support_scripts.utils.MastersModel import MastersModel
 from support_scripts.utils.ModelSettingsManager import ModelSettingsManager
 from support_scripts.utils.RunTimer import RunTimer
-from support_scripts.utils.custom_datasets import (
-    CityScapesDataset,
-    CityScapesDemoVideoDataset,
-)
+from support_scripts.utils.datasets.custom_datasets import CityScapesDataset
+from support_scripts.utils.datasets.custom_video_datasets import CityScapesDemoVideoDataset, CityScapesVideoDataset
 from support_scripts.utils.norm_management import norm_selector
+
+__all__ = [
+    "MastersModel",
+    "ModelSettingsManager",
+    "RunTimer",
+    "CityScapesVideoDataset",
+    "CityScapesDataset",
+    "CityScapesDemoVideoDataset",
+    "norm_selector",
+]
